Fix x86/x64 assumptions in code paths that ARM64 reaches
Various places in the tree assumed _MSC_VER implies x86 or that the legacy
WIN64 macro (only set manually in the .vcxproj for x64) tracks 64-bitness.
On Windows ARM64 those assumptions break - either at compile time
(intrinsics that don't exist) or at link time (symbols that are looked up
under x86-mangled names).

- 3rdparty/zstd/cpu.h: gate the __cpuid / __cpuidex MSVC intrinsics on
  _M_IX86 || _M_X64. zstd's CPU detection has a fallback path that simply
  reports no SIMD; ARM64 takes that.

- replay/version.cpp: the /include:GitVersionHash linker pragma was
  selected by `defined(WIN64)` vs `defined(WIN32)`. WIN64 is only defined
  manually in the .vcxproj for x64; ARM64 has _WIN64 (auto-set by the
  compiler for any 64-bit Windows target) but not WIN64, so the build was
  emitting /include:_GitVersionHash (the x86 mangled name) and failing to
  link. Switch to _WIN64 / _WIN32 which are universal.

- renderdocshim/renderdocshim.h: same fix for the GLOBAL_HOOK_DATA_NAME /
  SHIM_DLL_NAME selection.

- replay/renderdoc_serialise.inl: the SIZE_CHECK template asserts hardcode
  expected sizes for x64 layout. ARM64 padding/alignment can differ, so
  disable the development-only check on _M_ARM64.

- driver/vulkan/vk_win32.cpp: skip the Wow6432Node implicit-layer
  registration on ARM64 / ARM64EC. There's no x86 sibling DLL to register;
  Win11 ARM64's XTAJIT32 emulator is out of scope for phase 1.

- driver/ihv/amd/amd_isa_win32.cpp: cast HRESULT to int32_t before passing
  to ToStr(). HRESULT is `long` on Windows; serialiser.cpp has explicit
  DoStringise<int32_t> but not DoStringise<long>, so without the cast the
  template instantiation goes unresolved at link time. (This bug was latent
  on x64 too but happened to not be triggered there in our build config.)

+++ renderdoc/driver/ihv/amd/amd_isa_win32.cpp
@@ -300,8 +300,10 @@ rdcstr DisassembleDXBC(const bytebuf &shaderBytes, const rdcstr &target)
 
   if(out.pShaderBinary == NULL || out.shaderBinarySize < 16)
   {
+    // HRESULT is `long` on Windows; ToStr only has a specialisation for
+     // int32_t (which is `int`), not `long`, so cast to keep the linker happy.
     RDCLOG("Failed to disassemble shader: %p/%zu (%s)", out.pShaderBinary, out.shaderBinarySize,
-           ToStr(hr).c_str());
+           ToStr((int32_t)hr).c_str());
     return "; Failed to disassemble shader";
   }
